requirements.txt
python-telegram-bot[job-queue]>=20.0
pillow>=9.0.0
pystray>=0.19.0
